moving functionality to dataframe.py
